+++ Kawai.Domain/DTOs/ReceiptSupplyHistoryDto.cs
@@ -10,12 +10,12 @@ public class ReceiptSupplyHistoryDto: DataTableDto
     public string LotNo { get; set; }
     public string TransactionType { get; set; }
     public DateTime TransactionDate { get; set; }
-    public double PreMonth { get; set; }
-    public double Receipt { get; set; }
-    public double Supply { get; set; }
-    public double Reject { get; set; }
-    public double Current { get; set; }
-    public double QtyTrans { get; set; }
+    public decimal PreMonth { get; set; }
+    public decimal Receipt { get; set; }
+    public decimal Supply { get; set; }
+    public decimal Reject { get; set; }
+    public decimal Current { get; set; }
+    public decimal QtyTrans { get; set; }
     public string FromAreaName { get; set; }
     public string ToAreaName { get; set; }
     public string DocReference { get; set; }
